task full info
