+++ app/src/main/java/uz/androbeck/virtualbank/data/api/HomeService.kt
@@ -4,9 +4,7 @@ import retrofit2.http.GET
 import uz.androbeck.virtualbank.data.dto.request.FullInfoDto
 import uz.androbeck.virtualbank.utils.Constants
 
-interface FullInfoService {
-
+interface HomeService {
     @GET(Constants.Endpoint.FULL_INFO)
-    // Header() impl
-    suspend fun getFullInfo( authToken: String?): FullInfoDto
+    suspend fun getFullInfo(): FullInfoDto
 }
