+++ app/src/main/java/uz/androbeck/virtualbank/data/repository/authentication/AuthenticationRepository.kt
@@ -1,4 +1,4 @@
-package uz.androbeck.virtualbank.data.repository.authenticationRepository
+package uz.androbeck.virtualbank.data.repository.authentication
 
 import kotlinx.coroutines.flow.Flow
 import uz.androbeck.virtualbank.data.dto.common.response.TokenResDto
