+++ app/src/main/java/uz/androbeck/virtualbank/data/repository/authentication/AuthenticationRepositoryImpl.kt
@@ -1,4 +1,4 @@
-package uz.androbeck.virtualbank.data.repository.authenticationRepository
+package uz.androbeck.virtualbank.data.repository.authentication
 
 import uz.androbeck.virtualbank.data.dto.request.SignUpReqDto
 import uz.androbeck.virtualbank.data.source.remote.AuthenticationRemoteDataSource
