+++ app/src/main/java/uz/androbeck/virtualbank/data/repository/home/HomeRepository.kt
@@ -1,8 +1,8 @@
-package uz.androbeck.virtualbank.data.repository.mainRepository
+package uz.androbeck.virtualbank.data.repository.home
 
 import kotlinx.coroutines.flow.Flow
 import uz.androbeck.virtualbank.data.dto.request.FullInfoDto
 
 interface HomeRepository {
-    fun getFullInfo(token:String?):Flow<FullInfoDto>
+    fun getFullInfo():Flow<FullInfoDto>
 }
