+++ app/src/main/java/uz/androbeck/virtualbank/data/source/remote/HomeRemoteDataSource.kt
@@ -4,5 +4,5 @@ import kotlinx.coroutines.flow.Flow
 import uz.androbeck.virtualbank.data.dto.request.FullInfoDto
 
 interface HomeRemoteDataSource {
-    fun getFullInfo(token:String?):Flow<FullInfoDto>
+    fun getFullInfo():Flow<FullInfoDto>
 }
