+++ app/src/main/java/uz/androbeck/virtualbank/di/UseCaseModule.kt
@@ -4,13 +4,13 @@ import dagger.Module
 import dagger.Provides
 import dagger.hilt.InstallIn
 import dagger.hilt.android.components.ViewModelComponent
-import uz.androbeck.virtualbank.data.repository.authenticationRepository.AuthenticationRepository
-import uz.androbeck.virtualbank.data.repository.mainRepository.HomeRepository
+import uz.androbeck.virtualbank.data.repository.authentication.AuthenticationRepository
+import uz.androbeck.virtualbank.data.repository.home.HomeRepository
 import uz.androbeck.virtualbank.domain.mapper.auth.SignUpMapper
 import uz.androbeck.virtualbank.domain.mapper.auth.TokenMapper
-import uz.androbeck.virtualbank.domain.mapper.main.FullInfoMapper
+import uz.androbeck.virtualbank.domain.mapper.fullInfo.FullInfoMapper
 import uz.androbeck.virtualbank.domain.useCases.authentication.SignUpUseCase
-import uz.androbeck.virtualbank.domain.useCases.main.GetFullInfoUseCase
+import uz.androbeck.virtualbank.domain.useCases.home.GetFullInfoUseCase
 
 @Module
 @InstallIn(ViewModelComponent::class)
@@ -27,5 +27,5 @@ object UseCaseModule {
     fun provideMainUseCase(
         homeRepository: HomeRepository,
         fullInfoMapper: FullInfoMapper
-    ): GetFullInfoUseCase = GetFullInfoUseCase(homeRepository,fullInfoMapper)
+    ) = GetFullInfoUseCase(homeRepository,fullInfoMapper)
 }
