+++ app/src/main/java/uz/androbeck/virtualbank/domain/mapper/fullInfo/FullInfoMapper.kt
@@ -1,14 +1,14 @@
-package uz.androbeck.virtualbank.domain.mapper.main
+package uz.androbeck.virtualbank.domain.mapper.fullInfo
 
 import uz.androbeck.virtualbank.data.dto.request.FullInfoDto
 import uz.androbeck.virtualbank.domain.mapper.BaseMapper
-import uz.androbeck.virtualbank.domain.ui_models.main.FullInfoUIModel
+import uz.androbeck.virtualbank.domain.ui_models.home.FullInfoUIModel
 import javax.inject.Inject
 import javax.inject.Singleton
 
 @Singleton
 class FullInfoMapper @Inject constructor() : BaseMapper<FullInfoDto, FullInfoUIModel> {
-    override fun toUIModel(dto: FullInfoDto): FullInfoUIModel = dto.run {
+    override fun toUIModel(dto: FullInfoDto) = dto.run {
         FullInfoUIModel(
             firstName = first_name,
             lastName = last_name,
@@ -17,7 +17,7 @@ class FullInfoMapper @Inject constructor() : BaseMapper<FullInfoDto, FullInfoUIM
             gender = gender
         )
     }
-    override fun toDTO(uiModel: FullInfoUIModel): FullInfoDto = uiModel.run {
+    override fun toDTO(uiModel: FullInfoUIModel) = uiModel.run {
         FullInfoDto(
             first_name = firstName,
             last_name = lastName,
