+++ app/src/main/java/uz/androbeck/virtualbank/domain/ui_models/home/FullInfoUIModel.kt
@@ -1,4 +1,4 @@
-package uz.androbeck.virtualbank.domain.ui_models.main
+package uz.androbeck.virtualbank.domain.ui_models.home
 
 data class FullInfoUIModel(
     val phone: String? = null,
