+++ app/src/main/java/uz/androbeck/virtualbank/domain/useCases/authentication/SignUpUseCase.kt
@@ -2,7 +2,7 @@ package uz.androbeck.virtualbank.domain.useCases.authentication
 
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.map
-import uz.androbeck.virtualbank.data.repository.authenticationRepository.AuthenticationRepository
+import uz.androbeck.virtualbank.data.repository.authentication.AuthenticationRepository
 import uz.androbeck.virtualbank.domain.mapper.auth.SignUpMapper
 import uz.androbeck.virtualbank.domain.mapper.auth.TokenMapper
 import uz.androbeck.virtualbank.domain.ui_models.authentication.SignUpReqUIModel
