+++ app/src/main/java/uz/androbeck/virtualbank/ui/screens/bottom_nav_items/main/MainViewModel.kt
@@ -5,18 +5,16 @@ import androidx.lifecycle.viewModelScope
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.flow.launchIn
 import kotlinx.coroutines.flow.onEach
-import uz.androbeck.virtualbank.domain.useCases.main.GetFullInfoUseCase
-import uz.androbeck.virtualbank.preferences.PreferencesProvider
+import uz.androbeck.virtualbank.domain.useCases.home.GetFullInfoUseCase
 import javax.inject.Inject
 
 @HiltViewModel
 class MainViewModel @Inject constructor(
     private val getFullInfoUseCase: GetFullInfoUseCase,
-    private val provider: PreferencesProvider
 ) : ViewModel() {
 
     fun getUserFullInfo(){
-        getFullInfoUseCase.invoke(provider.token).onEach {
+        getFullInfoUseCase.invoke().onEach {
             println(it)
         }.launchIn(viewModelScope)
     }
